requirements.txt
Django
Pillow
django-filter
